% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_functions.R
\name{annotate_gene_cell_type_specificity_kpmp}
\alias{annotate_gene_cell_type_specificity_kpmp}
\title{Generate gene expression annotations}
\usage{
annotate_gene_cell_type_specificity_kpmp(gene_annotation)
}
\arguments{
\item{gene_annotation}{A list as output by 'annotate_genes'}
}
\value{
A list as output by 'annotate_genes', with gene expression annotation
artifact locations added
}
\description{
Build KPMP gene expression annotation artifacts
}
