% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_functions.R
\name{annotate_gene_maf}
\alias{annotate_gene_maf}
\title{Generate MAF tables}
\usage{
annotate_gene_maf(gene_annotation)
}
\arguments{
\item{gene_annotation}{A list as output by 'annotate_genes'}
}
\value{
A list as output by 'annotate_genes', with gene expression annotation
artifact locations added
}
\description{
Generate gene annotation artifacts for tissue-specific expression
}
