% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_functions.R
\name{annotate_genes}
\alias{annotate_genes}
\title{Annotate set of genes}
\usage{
annotate_genes(
  genelist,
  setname = "all",
  tissue = "all",
  data_dir,
  out_dir,
  edb = edb_Hs
)
}
\arguments{
\item{genelist}{A vector of Ensembl gene_ids}

\item{setname}{A character string (default "all") describing the gene grouping}

\item{tissue}{A character string (default "all") specifying the focal tissue}

\item{data_dir}{The location where output will be placed}

\item{edb}{An \code{edb} object, specific to species}
}
\value{
A list containing the annotation artifacts created, as files, in
\code{data_dir}
}
\description{
Set up initial gene annotation data structure
}
