% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_functions.R
\name{annotate_mouse_gene_phenotypes}
\alias{annotate_mouse_gene_phenotypes}
\title{Generate HTML table of mouse phenotypes by gene}
\usage{
annotate_mouse_gene_phenotypes(gene_annotation)
}
\arguments{
\item{gene}{The human gene of interest}
}
\value{
A formatted HTML table of phenotypes for the mouse ortholog
}
\description{
Build gene annotation artifacts for mouse genes
}
