% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation_functions.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate set of variants}
\usage{
annotate_variants(
  variantlist,
  setname = "all",
  tissue = "all",
  data_dir,
  edb = edb_Hs
)
}
\arguments{
\item{variantlist}{A tibble of variants from CKAN}

\item{setname}{A character string (default "all") describing the gene grouping}

\item{tissue}{A character string (default "all") specifying the focal tissue}

\item{data_dir}{The location where output will be placed}

\item{edb}{An \code{edb} object, specific to species}
}
\value{
A list containing the annotation artifacts created, as files, in
\code{data_dir}
}
\description{
Set up initial variant annotation data structure
}
