% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_publication_functions.R
\name{get_or_create_group_ids}
\alias{get_or_create_group_ids}
\title{Look for groups with given names. If a group with the name exists, return its
ID, otherwise create it and return its ID}
\usage{
get_or_create_group_ids(group_names, ckan_org, dry_run = TRUE)
}
\arguments{
\item{group_names}{Vector of names of potential groups}

\item{ckan_org}{CKAN organization object}
}
\value{
A named vector of group ids
}
\description{
Get a vector of group ID's, creating where missing
}
