% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_publication_functions.R
\name{get_or_create_package_id}
\alias{get_or_create_package_id}
\title{If a package with the given name exists, return its ID, otherwise create it
and return its ID}
\usage{
get_or_create_package_id(
  ckan_org_id,
  package_name,
  package_title,
  dry_run = dry_run
)
}
\description{
Get a package's ID or if not existing create and get ID
}
