% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_publication_functions.R
\name{publish_annotation}
\alias{publish_annotation}
\title{Publish annotations from a set of genes to CKAN}
\usage{
publish_annotation(
  annotation,
  ckan_org,
  update = TRUE,
  dry_run = FALSE,
  prefix = NULL
)
}
\arguments{
\item{annotation}{An annotation structure as returned by gene
annotation or variant annotation functions}

\item{ckan_org}{A list with the organization id and list of packages}

\item{update}{Whether to overwrite (default) or fail if resource exists}

\item{dry_run}{If TRUE, output publication steps without executing}
}
\value{
The status of the publication
}
\description{
Publish artifacts from annotation object to CKAN
}
